feat: setup dry-run mode

+++ packages/cli/src/bin/wuzzle-transpile/wuzzle-transpile.ts
@@ -9,27 +9,20 @@ import os from 'os';
 import pMap from 'p-map';
 import path from 'path';
 import rimraf from 'rimraf';
-import findUp from 'find-up';
 import type webpack from 'webpack';
 import {
   CHAR_CTRL_D,
   EK_COMMAND_ARGS,
   EK_COMMAND_NAME,
-  EK_PROJECT_PATH,
-  EK_RPOJECT_ANCHOR,
+  EK_DRY_RUN,
   EXIT_CODE_ERROR,
   EXIT_CODE_USER_TERMINATION,
 } from '../../constants';
 import { transpile } from '../../transpile';
+import { checkToUseDryRunMode, locateProjectAnchor } from '../../utils';
 
-const anchorName = process.env[EK_RPOJECT_ANCHOR] || 'package.json';
-const anchorPath = findUp.sync(anchorName);
-if (!anchorPath) {
-  console.error(`error: '${anchorName}' not located.`);
-  process.exit(EXIT_CODE_ERROR);
-}
-const projectPath = path.dirname(anchorPath);
-process.env[EK_PROJECT_PATH] = projectPath;
+locateProjectAnchor();
+checkToUseDryRunMode(process.argv);
 
 // Set command same env variables as 'wuzzle transpile'
 process.env[EK_COMMAND_NAME] = 'transpile';
@@ -159,6 +152,16 @@ async function launchExec() {
         : 'inline-cheap-module-source-map'
       : undefined;
 
+  const webpackConfig: webpack.Configuration = {
+    mode: webpackMode,
+    target: webpackTarget,
+    devtool: webpackDevtool,
+  };
+
+  if (process.env[EK_DRY_RUN]) {
+    return await transpile({ webpackConfig });
+  }
+
   // Check to clean output dir
   if (clean) {
     rimraf.sync(outDir);
@@ -172,15 +175,7 @@ async function launchExec() {
     inputPath = path.resolve(inputPath);
     const outputPath = path.join(outDir, path.relative(basePath, inputPath));
     try {
-      await transpile({
-        inputPath,
-        outputPath,
-        webpackConfig: {
-          mode: webpackMode,
-          target: webpackTarget,
-          devtool: webpackDevtool,
-        },
-      });
+      await transpile({ inputPath, outputPath, webpackConfig });
       if (inputPathsCompiled[inputPath]) {
         forceLog(grey(`File '${path.relative(process.cwd(), inputPath)}' recompiled.`));
       } else {
